feat: blade syntax highlighting
This fixes the issue where blade is not highlighted properly. The cause
of the issue is due to a small file name mistake `highlight` as used
instead of `highlights` that single letter caused the entire problem
but not totally as some parser needs to be installed as well namely
`html` and `php_only` both can be installed using `:TSInstall foo`.